Refactoring the datalayer
