+++ argos/datalayer/jsonDatalayer.py
@@ -1,0 +1,8 @@
+import json
+
+
+class fileDataLayerFactory:
+
+    def __init__(self,):
+        pass
+
